src/ml_service.py
"""Azure Machine Learning 연동 모듈"""
from typing import Optional, Dict, Any
from src.config import settings

# ML 관련 모듈 import는 선택사항
try:
    from azure.ai.ml import MLClient
    from azure.identity import DefaultAzureCredential
    ML_AVAILABLE = True
except ImportError:
    ML_AVAILABLE = False
    MLClient = None
    DefaultAzureCredential = None


class MLService:
    """Azure Machine Learning 서비스"""
    
    def __init__(self):
        """ML 클라이언트 초기화"""
        self.ml_client: Optional[Any] = None
        self._ml_available = False
        
        if not ML_AVAILABLE:
            print("ML 모듈을 사용할 수 없습니다 (선택사항)")
            return
        
        if all([
            settings.azure_ml_workspace_name,
            settings.azure_ml_resource_group,
            settings.azure_ml_subscription_id
        ]):
            try:
                credential = DefaultAzureCredential()
                self.ml_client = MLClient(
                    credential=credential,
                    subscription_id=settings.azure_ml_subscription_id,
                    resource_group_name=settings.azure_ml_resource_group,
                    workspace_name=settings.azure_ml_workspace_name,
                )
                self._ml_available = True
            except Exception as e:
                print(f"ML 클라이언트 초기화 실패: {e}")
    
    def predict(self, model_name: str, data: Dict[str, Any]) -> Optional[Dict]:
        """
        모델 예측 실행
        
        Args:
            model_name: 모델 이름
            data: 예측할 데이터
            
        Returns:
            예측 결과 또는 None
        """
        if not self.ml_client:
            return None
        
        try:
            # 엔드포인트를 통한 예측 (예시)
            # 실제 구현은 배포된 모델의 엔드포인트에 따라 달라집니다
            # endpoint = self.ml_client.online_endpoints.get(model_name)
            # result = endpoint.invoke(data)
            # return result
            
            # 임시 구현
            return {"prediction": "모델 예측 결과", "data": data}
        except Exception as e:
            print(f"예측 실패: {e}")
            return None
    
    def train_model(self, training_config: Dict[str, Any]) -> Optional[str]:
        """
        모델 학습 시작
        
        Args:
            training_config: 학습 설정
            
        Returns:
            작업 ID 또는 None
        """
        if not self.ml_client:
            return None
        
        try:
            # 학습 작업 시작 (예시)
            # job = self.ml_client.jobs.create_or_update(training_config)
            # return job.id
            
            # 임시 구현
            return "job_12345"
        except Exception as e:
            print(f"학습 실패: {e}")
            return None
